third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-goto-matching-bracket-expected.txt

This test verifies editor's "Goto Matching Bracket" behavior, which is triggered via Ctrl-M shortcut.

function MyClass(a, b)
{
    console.log("Test");
}


Running: testSingleCursorFromOutsideOpenBracket
function.MyClass(a,.b)|
{
....console.log("Test");
}


Running: testSingleCursorFromInsideOpenBracket
function.MyClass(a,.b|)
{
....console.log("Test");
}


Running: testSingleCursorFromOutsideCloseBracket
function.MyClass(a,.b)
|{
....console.log("Test");
}


Running: testSingleCursorFromInsideCloseBracket
function.MyClass(a,.b)
{|
....console.log("Test");
}


Running: testMulticursor
function.MyClass(|a,.b)|
{|
....consol|e.log("Test");
}
